WIP: rename home directory
