+++ apps/app/src/router/index.tsx
@@ -9,9 +9,9 @@ import { LoginPage } from '../routes/auth/login.page';
 import { ClassificationTasksRoute } from '../routes/classification-tasks';
 import { ClassificationTaskPage } from '../routes/classification-tasks/classification-task.page';
 import { ClassificationTasksPage } from '../routes/classification-tasks/classification-tasks.page';
-import { HomeRoute } from '../routes/home';
-import { DynamicHomePage } from '../routes/home/dynamic-home.page';
-import { HomePage } from '../routes/home/home.page';
+import { HomeRoute } from '../routes/home-old';
+import { DynamicHomePage } from '../routes/home-old/dynamic-home.page';
+import { HomePage } from '../routes/home-old/home.page';
 import { JobProfilesRoute } from '../routes/job-profiles';
 import { JobProfilesPage } from '../routes/job-profiles/job-profiles.page';
 import { SavedJobProfilesPage } from '../routes/job-profiles/saved-job-profiles.page';
