+++ apps/app/src/routes/home-old/components/initials-avatar.component.css
